Extract common superclasses for repeated readers
Creats a cross format superclass for repeated types. Does not change
the operation of the concerned DWRF readers.
