+++ velox/dwio/dwrf/reader/SelectiveRepeatedColumnReader.cpp
@@ -18,14 +18,36 @@
 #include "velox/dwio/dwrf/reader/SelectiveDwrfReader.h"
 
 namespace facebook::velox::dwrf {
+namespace {
+std::unique_ptr<dwio::common::IntDecoder</*isSigned*/ false>> makeLengthDecoder(
+    const dwio::common::TypeWithId& nodeType,
+    DwrfParams& params,
+    memory::MemoryPool& pool) {
+  EncodingKey encodingKey{nodeType.id, params.flatMapContext().sequence};
+  auto& stripe = params.stripeStreams();
+  auto rleVersion = convertRleVersion(stripe.getEncoding(encodingKey).kind());
+  auto lenId = encodingKey.forKind(proto::Stream_Kind_LENGTH);
+  bool lenVints = stripe.getUseVInts(lenId);
+  return createRleDecoder</*isSigned*/ false>(
+      stripe.getStream(lenId, true),
+      rleVersion,
+      pool,
+      lenVints,
+      dwio::common::INT_BYTE_SIZE);
+}
+} // namespace
 
 SelectiveListColumnReader::SelectiveListColumnReader(
     const std::shared_ptr<const dwio::common::TypeWithId>& requestedType,
     const std::shared_ptr<const dwio::common::TypeWithId>& dataType,
     DwrfParams& params,
     common::ScanSpec& scanSpec)
-    : SelectiveRepeatedColumnReader(dataType, params, scanSpec, dataType->type),
-      requestedType_{requestedType} {
+    : dwio::common::SelectiveListColumnReader(
+          requestedType,
+          dataType,
+          params,
+          scanSpec),
+      length_(makeLengthDecoder(*nodeType_, params, memoryPool_)) {
   DWIO_ENSURE_EQ(nodeType_->id, dataType->id, "working on the same node");
   EncodingKey encodingKey{nodeType_->id, params.flatMapContext().sequence};
   auto& stripe = params.stripeStreams();
@@ -47,69 +69,17 @@ SelectiveListColumnReader::SelectiveListColumnReader(
       childType, nodeType_->childAt(0), childParams, *scanSpec_->children()[0]);
 }
 
-uint64_t SelectiveListColumnReader::skip(uint64_t numValues) {
-  numValues = formatData_->skipNulls(numValues);
-  if (child_) {
-    std::array<int64_t, kBufferSize> buffer;
-    uint64_t childElements = 0;
-    uint64_t lengthsRead = 0;
-    while (lengthsRead < numValues) {
-      uint64_t chunk =
-          std::min(numValues - lengthsRead, static_cast<uint64_t>(kBufferSize));
-      length_->next(buffer.data(), chunk, nullptr);
-      for (size_t i = 0; i < chunk; ++i) {
-        childElements += static_cast<size_t>(buffer[i]);
-      }
-      lengthsRead += chunk;
-    }
-    child_->skip(childElements);
-    childTargetReadOffset_ += childElements;
-    child_->setReadOffset(child_->readOffset() + childElements);
-  } else {
-    length_->skip(numValues);
-  }
-  return numValues;
-}
-
-void SelectiveListColumnReader::read(
-    vector_size_t offset,
-    RowSet rows,
-    const uint64_t* incomingNulls) {
-  // Catch up if the child is behind the length stream.
-  child_->seekTo(childTargetReadOffset_, false);
-  prepareRead<char>(offset, rows, incomingNulls);
-  makeNestedRowSet(rows);
-  if (child_ && !nestedRows_.empty()) {
-    child_->read(child_->readOffset(), nestedRows_, nullptr);
-  }
-  numValues_ = rows.size();
-  readOffset_ = offset + rows.back() + 1;
-}
-
-void SelectiveListColumnReader::getValues(RowSet rows, VectorPtr* result) {
-  compactOffsets(rows);
-  VectorPtr elements;
-  if (child_ && !nestedRows_.empty()) {
-    prepareStructResult(type_->childAt(0), &elements);
-    child_->getValues(nestedRows_, &elements);
-  }
-  *result = std::make_shared<ArrayVector>(
-      &memoryPool_,
-      requestedType_->type,
-      anyNulls_ ? resultNulls_ : nullptr,
-      rows.size(),
-      offsets_,
-      sizes_,
-      elements);
-}
-
 SelectiveMapColumnReader::SelectiveMapColumnReader(
     const std::shared_ptr<const dwio::common::TypeWithId>& requestedType,
     const std::shared_ptr<const dwio::common::TypeWithId>& dataType,
     DwrfParams& params,
     common::ScanSpec& scanSpec)
-    : SelectiveRepeatedColumnReader(dataType, params, scanSpec, dataType->type),
-      requestedType_{requestedType} {
+    : dwio::common::SelectiveMapColumnReader(
+          requestedType,
+          dataType,
+          params,
+          scanSpec),
+      length_(makeLengthDecoder(*nodeType_, params, memoryPool_)) {
   DWIO_ENSURE_EQ(nodeType_->id, dataType->id, "working on the same node");
   EncodingKey encodingKey{nodeType_->id, params.flatMapContext().sequence};
   auto& stripe = params.stripeStreams();
@@ -148,84 +118,6 @@ SelectiveMapColumnReader::SelectiveMapColumnReader(
       *scanSpec_->children()[1]);
 
   VLOG(1) << "[Map] Initialized map column reader for node " << nodeType_->id;
-}
-
-uint64_t SelectiveMapColumnReader::skip(uint64_t numValues) {
-  numValues = formatData_->skipNulls(numValues);
-  if (keyReader_ || elementReader_) {
-    std::array<int64_t, kBufferSize> buffer;
-    uint64_t childElements = 0;
-    uint64_t lengthsRead = 0;
-    while (lengthsRead < numValues) {
-      uint64_t chunk =
-          std::min(numValues - lengthsRead, static_cast<uint64_t>(kBufferSize));
-      length_->next(buffer.data(), chunk, nullptr);
-      for (size_t i = 0; i < chunk; ++i) {
-        childElements += buffer[i];
-      }
-      lengthsRead += chunk;
-    }
-    if (keyReader_) {
-      keyReader_->skip(childElements);
-      keyReader_->setReadOffset(keyReader_->readOffset() + childElements);
-    }
-    if (elementReader_) {
-      elementReader_->skip(childElements);
-      elementReader_->setReadOffset(
-          elementReader_->readOffset() + childElements);
-    }
-    childTargetReadOffset_ += childElements;
-
-  } else {
-    length_->skip(numValues);
-  }
-  return numValues;
-}
-
-void SelectiveMapColumnReader::read(
-    vector_size_t offset,
-    RowSet rows,
-    const uint64_t* incomingNulls) {
-  // Catch up if child readers are behind the length stream.
-  if (keyReader_) {
-    keyReader_->seekTo(childTargetReadOffset_, false);
-  }
-  if (elementReader_) {
-    elementReader_->seekTo(childTargetReadOffset_, false);
-  }
-
-  prepareRead<char>(offset, rows, incomingNulls);
-  makeNestedRowSet(rows);
-  if (keyReader_ && elementReader_ && !nestedRows_.empty()) {
-    keyReader_->read(keyReader_->readOffset(), nestedRows_, nullptr);
-    elementReader_->read(elementReader_->readOffset(), nestedRows_, nullptr);
-  }
-  numValues_ = rows.size();
-  readOffset_ = offset + rows.back() + 1;
-}
-
-void SelectiveMapColumnReader::getValues(RowSet rows, VectorPtr* result) {
-  compactOffsets(rows);
-  VectorPtr keys;
-  VectorPtr values;
-  VELOX_CHECK(
-      keyReader_ && elementReader_,
-      "keyReader_ and elementReaer_ must exist in "
-      "SelectiveMapColumnReader::getValues");
-  if (!nestedRows_.empty()) {
-    keyReader_->getValues(nestedRows_, &keys);
-    prepareStructResult(type_->childAt(1), &values);
-    elementReader_->getValues(nestedRows_, &values);
-  }
-  *result = std::make_shared<MapVector>(
-      &memoryPool_,
-      requestedType_->type,
-      anyNulls_ ? resultNulls_ : nullptr,
-      rows.size(),
-      offsets_,
-      sizes_,
-      keys,
-      values);
 }
 
 } // namespace facebook::velox::dwrf
